combine derivative creation in single file, store PDFs in separate dir
